+++ src/galeforce/actions/tft/tft-league/leagues.ts
@@ -20,7 +20,7 @@ export default class GetTFTLeagueList extends BaseAction<LeagueListDTO> {
         this.payload.method = 'GET';
     }
 
-    public async exec(): Promise<LeagueListDTO> {
+    protected inferEndpoint(): void {
         if (this.payload.leagueId) {
             this.payload.endpoint = ENDPOINTS.TFT_LEAGUE.LEAGUE_ID;
         } else if (this.payload.tier) {
@@ -40,7 +40,5 @@ export default class GetTFTLeagueList extends BaseAction<LeagueListDTO> {
         } else {
             throw new Error('[galeforce]: Not enough parameters provided to select API endpoint.');
         }
-
-        return super.exec();
     }
 }